requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
asdl==0.1.5
astor==0.7.1
attrs==18.2.0
audioread==2.1.9
Babel==2.7.0
bitarray==2.4.1
blessings==1.7
boto3==1.21.35
botocore==1.24.35
bpemb==0.2.12
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
corenlp-protobuf==3.8.0
Cython==0.29.28
decorator==5.1.1
docopt==0.6.2
entmax==1.0
et-xmlfile==1.1.0
gensim==4.1.2
google-auth==2.6.6
google-auth-oauthlib==0.4.6
gpustat==0.6.0
greenlet==1.1.2
grpcio==1.44.0
hydra-core==1.0.7
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
jdcal==1.4.1
jmespath==1.0.0
joblib==1.1.0
jsonnet==0.14.0
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
more-itertools==8.12.0
munch==2.5.0
networkx==2.6.3
nltk==3.7
num2words==0.5.10
numba==0.55.1
numpy==1.21.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.0.6
openpyxl==2.4.11
packaging==21.3
pluggy==0.13.1
pooch==1.6.0
portalocker==2.4.0
protobuf==3.20.0
psutil==5.9.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
pyrsistent==0.14.11
pytest==5.3.5
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
records==0.5.3
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
s3transfer==0.5.2
sacrebleu==2.0.0
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
SoundFile==0.10.3.post1
sox==1.4.1
speechpy==2.4
SQLAlchemy==1.4.35
stanford-corenlp==3.9.2
stanfordcorenlp==3.9.1.1
tablib==3.2.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchtext==0.3.1
tqdm==4.36.1
transformers==2.3.0
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
